worker's rating
+ Fix worker image bug
+ Enhance ui

+++ gps_tracking_system/lib/Utility/RestApi/admin_get_worker_rating_response.dart
@@ -17,7 +17,7 @@ class GetWorkerRatingResponse {
 
   factory GetWorkerRatingResponse.fromJson(Map<String, dynamic> json) => GetWorkerRatingResponse(
     ratingList: List<RatingList>.from(json["rating_list"].map((x) => RatingList.fromJson(x))),
-    averageRating: json["average_rating"],
+    averageRating: json["average_rating"].toDouble(),
     response: Response.fromJson(json["response"]),
   );
